+++ docs/source/changelog.rst
@@ -6,11 +6,13 @@ Version 3.0.0
 
 .. versionadded:: 3.0.0
 
-   * :class:`.modules.voxels`'s, providing support for deep learning on voxel
-     data.
+   * :class:`.transforms.voxels`'s, providing support for transforming voxels.
+   * :class:`.modules.voxels`'s, providing support for deep learning on voxels.
 
 .. versionchanged:: 3.0.0
 
+    * Split :mod:`.tranforms` into :mod:`.transforms.points` and
+      :mod:`.transforms.voxels`.
     * Split :mod:`.modules` into :mod:`.modules.points` and
       :mod:`.modules.voxels`.
     * Renamed :class:`.PCDDataset` to :class:`.Dataset`.